refactor: アニメ関係をコンポーネント分離する
