+++ apps/mobile/components/anime-list/animeListStyles.ts
@@ -1,0 +1,419 @@
+import { StyleSheet } from "react-native";
+import { GRID_GAP } from "./animeListUtils";
+
+export const styles = StyleSheet.create({
+  screen: {
+    flex: 1,
+    backgroundColor: "#f6f3ee",
+  },
+  listContent: {
+    rowGap: 12,
+  },
+  columnWrapper: {
+    gap: GRID_GAP,
+    justifyContent: "flex-start",
+    marginBottom: GRID_GAP,
+  },
+  header: {
+    gap: 16,
+    marginBottom: 2,
+  },
+  hero: {
+    overflow: "hidden",
+    borderRadius: 8,
+    backgroundColor: "#fffaf2",
+    borderWidth: 1,
+    borderColor: "#efe4d4",
+    padding: 18,
+    shadowColor: "#111827",
+    shadowOffset: { width: 0, height: 12 },
+    shadowOpacity: 0.08,
+    shadowRadius: 24,
+    elevation: 3,
+  },
+  heroWide: {
+    padding: 28,
+  },
+  heroAccent: {
+    position: "absolute",
+    top: 0,
+    left: 0,
+    right: 0,
+    height: 5,
+    backgroundColor: "#eab308",
+  },
+  heroTop: {
+    gap: 18,
+  },
+  heroTopWide: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between",
+    gap: 32,
+  },
+  heroCopy: {
+    flex: 1,
+    minWidth: 0,
+  },
+  heroControls: {
+    flex: 1,
+    minWidth: 0,
+    width: "100%",
+  },
+  eyebrow: {
+    color: "#7c2d12",
+    fontSize: 11,
+    fontWeight: "800",
+    letterSpacing: 0,
+  },
+  title: {
+    marginTop: 5,
+    color: "#111827",
+    fontSize: 28,
+    lineHeight: 34,
+    fontWeight: "900",
+    letterSpacing: 0,
+  },
+  titleWide: {
+    fontSize: 38,
+    lineHeight: 44,
+  },
+  subtitle: {
+    marginTop: 7,
+    color: "#64748b",
+    fontSize: 13,
+    lineHeight: 19,
+    fontWeight: "500",
+  },
+  subtitleWide: {
+    maxWidth: 520,
+    fontSize: 15,
+    lineHeight: 22,
+  },
+  searchBar: {
+    minHeight: 50,
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 10,
+    borderRadius: 8,
+    backgroundColor: "#ffffff",
+    borderWidth: 1,
+    borderColor: "#e2e8f0",
+    paddingHorizontal: 13,
+  },
+  searchInput: {
+    flex: 1,
+    minWidth: 0,
+    color: "#111827",
+    fontSize: 15,
+    fontWeight: "600",
+    paddingVertical: 13,
+  },
+  clearButton: {
+    width: 30,
+    height: 30,
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 15,
+    backgroundColor: "#f1f5f9",
+  },
+  statsRow: {
+    marginTop: 13,
+    flexDirection: "row",
+    justifyContent: "space-between",
+    width: "100%",
+  },
+  statPill: {
+    width: "31.5%",
+    minWidth: 0,
+    borderRadius: 8,
+    borderWidth: 1,
+    paddingHorizontal: 10,
+    paddingVertical: 9,
+  },
+  statPill_ink: {
+    backgroundColor: "#f8fafc",
+    borderColor: "#e2e8f0",
+  },
+  statPill_gold: {
+    backgroundColor: "#fffbeb",
+    borderColor: "#fde68a",
+  },
+  statPill_mint: {
+    backgroundColor: "#ecfdf5",
+    borderColor: "#bbf7d0",
+  },
+  statValue: {
+    color: "#111827",
+    fontSize: 14,
+    fontWeight: "900",
+  },
+  statLabel: {
+    marginTop: 1,
+    color: "#64748b",
+    fontSize: 10,
+    fontWeight: "700",
+  },
+  toolbar: {
+    gap: 10,
+  },
+  toolbarWide: {
+    flexDirection: "row",
+    alignItems: "flex-end",
+    justifyContent: "space-between",
+  },
+  sectionLabel: {
+    color: "#7c2d12",
+    fontSize: 11,
+    fontWeight: "800",
+    letterSpacing: 0,
+  },
+  resultText: {
+    marginTop: 2,
+    color: "#111827",
+    fontSize: 17,
+    lineHeight: 22,
+    fontWeight: "900",
+  },
+  sortGroup: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    width: "100%",
+  },
+  sortGroupWide: {
+    width: 360,
+  },
+  sortButton: {
+    width: "49%",
+    minHeight: 36,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    gap: 5,
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "#dbe4ee",
+    backgroundColor: "#ffffff",
+    paddingHorizontal: 10,
+  },
+  sortButtonActive: {
+    borderColor: "#111827",
+    backgroundColor: "#ffffff",
+  },
+  sortText: {
+    color: "#64748b",
+    fontSize: 12,
+    fontWeight: "800",
+  },
+  sortTextActive: {
+    color: "#111827",
+  },
+  card: {
+    flexDirection: "row",
+    gap: 13,
+    borderRadius: 8,
+    backgroundColor: "#ffffff",
+    borderWidth: 1,
+    borderColor: "#e5e7eb",
+    padding: 10,
+    shadowColor: "#111827",
+    shadowOffset: { width: 0, height: 8 },
+    shadowOpacity: 0.05,
+    shadowRadius: 18,
+    elevation: 2,
+  },
+  poster: {
+    width: 68,
+    height: 92,
+    borderRadius: 6,
+    backgroundColor: "#e2e8f0",
+  },
+  posterPlaceholder: {
+    width: 68,
+    height: 92,
+    alignItems: "center",
+    justifyContent: "center",
+    overflow: "hidden",
+    borderRadius: 6,
+    borderWidth: 1,
+  },
+  posterPlaceholderAccent: {
+    position: "absolute",
+    top: 0,
+    left: 0,
+    right: 0,
+    height: 4,
+  },
+  posterInitial: {
+    fontSize: 25,
+    lineHeight: 30,
+    fontWeight: "900",
+  },
+  posterFallbackLabel: {
+    marginTop: 2,
+    fontSize: 8,
+    fontWeight: "900",
+  },
+  cardBody: {
+    flex: 1,
+    minWidth: 0,
+    justifyContent: "center",
+  },
+  cardTitleRow: {
+    flexDirection: "row",
+    alignItems: "flex-start",
+    gap: 8,
+  },
+  cardTitle: {
+    flex: 1,
+    minWidth: 0,
+    color: "#111827",
+    fontSize: 16,
+    lineHeight: 21,
+    fontWeight: "900",
+    letterSpacing: 0,
+  },
+  englishTitle: {
+    marginTop: 4,
+    color: "#64748b",
+    fontSize: 12,
+    lineHeight: 16,
+    fontWeight: "600",
+  },
+  metaRow: {
+    marginTop: 10,
+    flexDirection: "row",
+    alignItems: "center",
+    flexWrap: "wrap",
+    gap: 6,
+  },
+  seasonChip: {
+    minHeight: 25,
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 4,
+    borderRadius: 7,
+    backgroundColor: "#fef3c7",
+    paddingHorizontal: 8,
+  },
+  seasonText: {
+    color: "#92400e",
+    fontSize: 11,
+    fontWeight: "900",
+  },
+  genreChip: {
+    maxWidth: 92,
+    minHeight: 25,
+    justifyContent: "center",
+    borderRadius: 7,
+    backgroundColor: "#f1f5f9",
+    paddingHorizontal: 8,
+  },
+  genreText: {
+    color: "#475569",
+    fontSize: 11,
+    fontWeight: "800",
+  },
+  moreGenres: {
+    color: "#64748b",
+    fontSize: 11,
+    fontWeight: "900",
+  },
+  favoriteButton: {
+    width: 36,
+    height: 36,
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 18,
+    backgroundColor: "#f8fafc",
+    borderWidth: 1,
+    borderColor: "#e2e8f0",
+  },
+  favoriteButtonActive: {
+    backgroundColor: "#fff1f2",
+    borderColor: "#fecdd3",
+  },
+  emptyState: {
+    alignItems: "center",
+    justifyContent: "center",
+    paddingHorizontal: 30,
+    paddingVertical: 54,
+  },
+  stateScreen: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: "#f6f3ee",
+    paddingHorizontal: 32,
+  },
+  loadingMark: {
+    width: 56,
+    height: 56,
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 8,
+    backgroundColor: "#ffffff",
+    borderWidth: 1,
+    borderColor: "#e5e7eb",
+  },
+  stateIcon: {
+    width: 56,
+    height: 56,
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 8,
+    backgroundColor: "#ffffff",
+    borderWidth: 1,
+    borderColor: "#e5e7eb",
+  },
+  errorIcon: {
+    backgroundColor: "#fff1f2",
+    borderColor: "#fecdd3",
+  },
+  stateTitle: {
+    marginTop: 16,
+    color: "#111827",
+    textAlign: "center",
+    fontSize: 19,
+    lineHeight: 25,
+    fontWeight: "900",
+  },
+  stateCopy: {
+    marginTop: 7,
+    color: "#64748b",
+    textAlign: "center",
+    fontSize: 13,
+    lineHeight: 20,
+    fontWeight: "600",
+  },
+  primaryButton: {
+    marginTop: 20,
+    minHeight: 46,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    gap: 8,
+    borderRadius: 8,
+    backgroundColor: "#111827",
+    paddingHorizontal: 20,
+  },
+  primaryButtonText: {
+    color: "#ffffff",
+    fontSize: 14,
+    fontWeight: "900",
+  },
+  emptyTitle: {
+    marginTop: 14,
+    color: "#111827",
+    fontSize: 18,
+    fontWeight: "900",
+  },
+  emptyCopy: {
+    marginTop: 7,
+    color: "#64748b",
+    textAlign: "center",
+    fontSize: 13,
+    lineHeight: 20,
+    fontWeight: "600",
+  },
+});
