+++ apps/mobile/components/anime-list/index.ts
@@ -1,0 +1,1 @@
+export { AnimeListContent } from "./AnimeListContent";
